+++ src/Robot.cpp
@@ -232,15 +232,15 @@ void Robot::UpdateClaw(pros::Controller& master)
 void Robot::UpdateDrive(pros::Controller& master)
 {
 	// Arcade drive mode
-	float leftDrivePower = master.get_analog(E_CONTROLLER_ANALOG_LEFT_Y) 
+	double leftDrivePower = master.get_analog(E_CONTROLLER_ANALOG_LEFT_Y) 
 						+ master.get_analog(E_CONTROLLER_ANALOG_RIGHT_X);
-	float rightDrivePower = master.get_analog(E_CONTROLLER_ANALOG_LEFT_Y)
+	double rightDrivePower = master.get_analog(E_CONTROLLER_ANALOG_LEFT_Y)
 						- master.get_analog(E_CONTROLLER_ANALOG_RIGHT_X);
 
 	// Tank drive mode
 	/*
-	float leftDrivePower = master.get_analog(E_CONTROLLER_ANALOG_LEFT_Y);
-	float rightDrivePower = master.get_analog(E_CONTROLLER_ANALOG_RIGHT_Y);
+	double leftDrivePower = master.get_analog(E_CONTROLLER_ANALOG_LEFT_Y);
+	double rightDrivePower = master.get_analog(E_CONTROLLER_ANALOG_RIGHT_Y);
 	*/
 
 	// Update
